CHANGE table_output to general_tools
