+++ settings.py
@@ -15,3 +15,5 @@ LISA_URL = config('LISA')
 RANKED_SPREADSHEET_ID = '1E2cQBWeQc9JkCKv3BUPClGwulPXXg-4hTYotuUKmoJI'
 SCORE_INDEX = 4
 SD_NAME_INDEX = 1
+COLOR_INDEX = 1
+ELO_IMG_INDEX = 2
